Add hidden config option for smoothswitching count

+++ Source/CPController.h
@@ -21,7 +21,7 @@
 	NSString *currentContextUUID, *currentContextName;
 	NSString *guessConfidence;
 	BOOL guessIsConfident;
-	int smoothCounter;
+	NSInteger smoothCounter;
 
 	IBOutlet NSMenuItem *forceContextMenuItem;
 	BOOL forcedContextIsSticky;
